[refactor] Moved test for warning generated when sync functions are marked with "asyncio" into a separate test module.
The "test_simple" module is named too general. Thus, it serves as a magnet for all kinds of tests that aren't connected to each other. This is one step to break up the "test_simple" module into more coherent test modules.
